packages/auto-browser/controller/requirements.txt
fastapi==0.120.4
starlette==0.49.3
uvicorn[standard]==0.35.0
playwright==1.56.0
pydantic-settings==2.10.1
httpx==0.28.1
redis==7.1.1
cryptography==46.0.7
Pillow==12.2.0
pytesseract==0.3.13
docker==7.1.0
prometheus-client==0.22.1
pyotp==2.9.0
apscheduler==3.11.0
tomli==2.3.0; python_version < "3.11"
